make all the model editing views safe

+++ policykit/policyengine/views.py
@@ -165,7 +165,7 @@ def disable_integration(request):
     """
     # name of the plugin
     name = request.GET.get("name")
-    # id of the plugin (for disabling only)
+    # id of the plugin
     id = request.GET.get("id")
     assert id is not None and name is not None
 
@@ -184,6 +184,16 @@ def disable_integration(request):
     # We would need to delete the SlackCommunity as well, which we should show a warning for!
     if name == "slack":
         return redirect("/main/settings?error=cant_delete_slack")
+
+    # Validate that this plugin ID is valid for the community that this user is logged into.
+    # Important! This prevents the user from disabling plugins for other communities.
+    plugins = MetagovAPI.get_metagov_community(community.community.metagov_slug)["plugins"]
+    plugin_found = False
+    for p in plugins:
+        if p["name"] == name and p["id"] == id:
+            plugin_found = True
+    if not plugin_found:
+        return redirect("/main/settings?error=no_such_plugin")
 
     logger.debug(f"Deleting plugin {name} {id}")
     MetagovAPI.delete_plugin(name=name, id=id)
@@ -218,10 +228,8 @@ def editor(request):
         from policyengine.models import Policy
         policy = None
         try:
-            policy = Policy.objects.get(id=policy_id)
+            policy = Policy.objects.get(id=policy_id, community=user.community.community)
         except Policy.DoesNotExist:
-            return HttpResponseNotFound()
-        if policy.community != user.community.community:
             return HttpResponseNotFound()
 
         data['policy'] = policy_id
@@ -291,7 +299,7 @@ def roleeditor(request):
     }
 
     if role_pk:
-        role = CommunityRole.objects.get(pk=role_pk)
+        role = CommunityRole.objects.get(pk=role_pk, community=user.community.community)
         data['role_name'] = role.role_name
         data['name'] = role.name
         data['description'] = role.description
@@ -361,7 +369,11 @@ def documenteditor(request):
     }
 
     if doc_id:
-        doc = CommunityDoc.objects.filter(id=doc_id)[0]
+        try:
+            doc = CommunityDoc.objects.get(id=doc_id, community=user.community.community)
+        except CommunityDoc.DoesNotExist:
+            return HttpResponseNotFound()
+
         data['name'] = doc.name
         data['text'] = doc.text
 
@@ -499,10 +511,8 @@ def policy_action_save(request):
             action = PolicykitChangeTriggerPolicy()
         
         try:
-            action.policy = Policy.objects.get(pk=data['policy'])
+            action.policy = Policy.objects.get(pk=data['policy'], community=user.community.community)
         except Policy.DoesNotExist:
-            return HttpResponseNotFound()
-        if action.policy.community != user.community.community:
             return HttpResponseNotFound()
 
     else:
@@ -554,10 +564,8 @@ def policy_action_remove(request):
 
     action = None
     try:
-        policy = Policy.objects.get(pk=data['policy'])
+        policy = Policy.objects.get(pk=data['policy'], community=user.community.community)
     except Policy.DoesNotExist:
-        return HttpResponseNotFound()
-    if policy.community != user.community.community:
         return HttpResponseNotFound()
 
     if policy.kind == Policy.CONSTITUTION:
@@ -585,10 +593,8 @@ def policy_action_recover(request):
 
     action = None
     try:
-        policy = Policy.objects.get(pk=data['policy'])
+        policy = Policy.objects.get(pk=data['policy'], community=user.community.community)
     except Policy.DoesNotExist:
-        return HttpResponseNotFound()
-    if policy.community != user.community.community:
         return HttpResponseNotFound()
 
     if policy.kind == Policy.CONSTITUTION:
@@ -670,7 +676,7 @@ def role_action_remove(request):
     action = PolicykitDeleteRole()
     action.community = user.constitution_community
     action.initiator = user
-    action.role = CommunityRole.objects.get(pk=data['role'])
+    action.role = CommunityRole.objects.get(pk=data['role'], community=user.community.community)
     action.save()
 
     return HttpResponse()
@@ -711,7 +717,7 @@ def document_action_remove(request):
     action = PolicykitDeleteCommunityDoc()
     action.community = user.constitution_community
     action.initiator = user
-    action.doc = CommunityDoc.objects.get(id=data['doc'])
+    action.doc = CommunityDoc.objects.get(id=data['doc'], community=user.community.community)
     action.save()
 
     return HttpResponse()
@@ -727,7 +733,7 @@ def document_action_recover(request):
     action = PolicykitRecoverCommunityDoc()
     action.community = user.constitution_community
     action.initiator = user
-    action.doc = CommunityDoc.objects.get(id=data['doc'])
+    action.doc = CommunityDoc.objects.get(id=data['doc'], community=user.community.community)
     action.save()
 
     return HttpResponse()
